Move value type code into its own module. Fix udf test failures that popped up
